Bug fixes
Crash if a player escort about to enter a stargate with the player, had
a target, and that target entered a stargate. The problem was that we
assumed that the target was another escort and didn't clear out
m_pTarget.

This also fixes a bug with maneuvering code. IsPointingTo was returning
false incorrectly.