Fixed bug with radius explosions
Due to various changes, we cannot reliably determine the lifetime of an
effect while loading a weaponDesc XML (we don't know stuff until bind).

Because of that, the code that tried to set the lifetime of radius
weapons was busted. Instead, we set the radius lifetime at runtime (when
the weapon is fired) instead of at load time.