Fixed objGetItemProperty returning incorrectly for AutoDefenseDevices for certain properties (such as 'enabled and 'external).
objGetItemProperty on AutoDefenseDevices, for all properties except for fireRate and fireDelay, will look first at the weapon
that the AutoDefenseDevice is using, then check the device if the property is not valid for the weapon. This means that the
'external and 'enabled properties are taken from the weapon rather than the AutoDefenseDevice, and in the case of the latter
always returns Nil as the weapon is not actually installed on the ship. This fix causes objGetItemProperty to check the
device rather than the weapon (as is done for fireRate and fireDelay) for the 'external and 'enabled properties.