Added <OnArmorConsumePower> event to armor.
This event allows users to specify custom behavior on armor that causes
power consumption, as well as customized armor power usage. To enable
this, set the armor's "useOnArmorConsumePower" attribute to True.

If the event is enabled, then power equal to powerUse will be consumed
if the event returns True. If the returned value is an integer, that
amount of power will be consumed.