Increase shot lifetime limit for AI firing.
While deliberate, this is the cause of the commonly mentioned "AI range bug", making them unable to fire at long range.

Ideally, they might use the full range only if necessary to hit a target they can't close on. But the greater number of facings for most ships now makes it more reasonable to try to hit at long range.