round should return an integer
For CC, there is no advantage in returning a double (and I don't think there is any downside, since functions will coerce the type appropriately). The only possible problem is loss of range (e.g., double numbers can be much larger than max int. We can fix that in the future by testing for size and returning a double if necessary.