Consume ammo at the beginning of a repeating shot
This fixes a bug in which the fire rate for a weapon is under the repeating count. In that case, we never consume ammo (because we used to consume ammo only at the end of a burst).

I'm not sure why we originally decided to consume ammo at the end. Looking back on the code, this decision looks pretty old.