Fix undefined behaviour with textures on primary canvas
Note, for our render queue we should have a limit on the max number of
textures. We can generate as many textures as we want, and probably bind
them too, it's just the maximum number of texture units we can have
active. Just don't run out of memory...

We should keep all of the lazyloaded textures in memory just as vanilla
does. We don't have to update them after all!