Fix `burstTracksTargets` repeating weapon aim when target destroyed during burst
Presently, when a gun with `burstTracksTargets` is aiming at a target that
is destroyed during the burst, the gun continues the burst at that given
firing angle regardless of the ship's orientation (even if the firing
angle is outside the gun's fire arc). This fix makes the gun fire down the
default fire angle if the target is destroyed during the burst.