Fix save file corruption bug
This will not restore old save files, but it will make sure new files are saved correctly.

The bug was introduced in this commit: https://github.com/kronosaur/TranscendenceDev/commit/bb40cbe1a2a954f6eaa787e90cf9399c7a4aef26

In COrderList::WriteToStream we threw if pSystem was NULL. But pSystem is NULL for ascended objects, so we ended up crashing in the middle of saving. The exception was caught, so we continued saving but there was missing data.