Fix LRS icon for lootable stations (again)  (#76)
* Replace (scrTranslate gScreen...) with (typTranslate gType...) in Text elements

in case the type is asked to translate them while it isn't the active screen.

* scrTranslate gScreen -> typTranslate gType in Text elements

* Fix LRS icon for lootable stations (again)

https://ministry.kronosaur.com/record.hexm?id=61069

This logic is a little more robust as it doesn't affect defense turrets, walls, nav signs, barricades, and slave quarters (since they're immobile), or mining pods and chimera cocoons (no docking ports, though it might be nice to be able to recover pods).

I made the log buoy mobile and removed immutable from flotsam and the Dall sphere so it will affect them. This also restores the ability to dock with the last two, which was broken either by my last change or an engine change since then. Having no structural hp still makes them indestructible.

* Fix ability to recover barricades that were not deployed manually

* Fix merge conflict better

* Formatting